создаст директорию mocks с данными для приложения
