+++ src/components/filter.js
@@ -1,17 +1,17 @@
 export const getFilterTemplate = (filters) => {
   return `
     <section class="main__filter filter container"> 
-      ${filters.map((fiter) => `
+      ${filters.map(({title, count}) => `
         <input
           type="radio"
-          id="filter__${fiter.title}"
+          id="filter__${title}"
           class="filter__input visually-hidden"
           name="filter"
           checked
         />
-        <label for="filter__${fiter.title}" class="filter__label">
-        ${fiter.title}
-        <span class="filter__${fiter.title}-count">${fiter.count}</span></label
+        <label for="filter__${title}" class="filter__label">
+        ${title}
+        <span class="filter__${title}-count">${count}</span></label
         >
       `).join(``)}
     </section>`.trim();
