+++ src/main.js
@@ -6,7 +6,8 @@ import {getSortingTemplate} from './components/sorting.js';
 import {getTaskFormTemplate} from './components/task-form.js';
 import {getTaskTemplate} from './components/task.js';
 import {getLoadMoreTemplate} from './components/load-more.js';
-import {getTasksList, getFilter} from './data.js';
+import filter from './mocks/filter.js';
+import tasks from './mocks/task.js';
 
 const render = (wrap, template, place) => {
   wrap.insertAdjacentHTML(place, template);
@@ -17,7 +18,7 @@ const menu = document.querySelector(`.main__control`);
 
 render(menu, getMenuTemplate(), `beforeEnd`);
 render(main, getSearchTemplate(), `beforeEnd`);
-render(main, getFilterTemplate(getFilter()), `beforeEnd`);
+render(main, getFilterTemplate(filter), `beforeEnd`);
 render(main, getTasksContainerTemplate(), `beforeEnd`);
 
 const tasksContainer = document.querySelector(`.board`);
@@ -25,7 +26,7 @@ const tasksList = document.querySelector(`.board__tasks`);
 
 render(tasksList, getSortingTemplate(), `beforeBegin`);
 
-const tasks = getTasksList();
+
 const renderTasksListRange = (list, from, to) => {
   if (list.length <= to) {
     to = list.length;
