+++ src/mocks/filter.js
@@ -1,0 +1,30 @@
+export default [
+    {
+      title: `All`,
+      count: 0
+    },
+    {
+      title: `Overdue`,
+      count: 0
+    },
+    {
+      title: `Today`,
+      count: 3
+    },
+    {
+      title: `Favorites`,
+      count: 10
+    },
+    {
+      title: `Repeating`,
+      count: 0
+    },
+    {
+      title: `Tags`,
+      count: 0
+    },
+    {
+      title: `Archive`,
+      count: 20
+    }
+]
